Add hermetic built pkgconfig toolchain
The registration of the preinstalled toolchains have been moved to the
preinstalled_toolchains macro, so that the hermetic pkgconfig toolchain
is the default.

As the hermetic toolchains are registered first, there is no need to
specify register_preinstalled_toolchains=False in order for the build to
be as hermetic as possible.