Add toolchain executables to the PATH
Previously, some toolchain executables, e.g. make, pkg-config etc. were
not added to the PATH correctly. As such, even if make and pkg-config
were built from source, the system executables were used.

This commit modifies the tools_files attribute of the rules_foreign_cc
framework to instead be "tools_data", which provides path to the
executable and the target that provides it.